refactor : SSR fetch함수 분리
